chore: refactoring RFC Document
